feat(condo): DOMA-6556 Guard was moved to mutation with resident creating
