+++ apps/condo/domains/resident/constants/constants.js
@@ -163,7 +163,7 @@ const PAYMENT_CATEGORIES_META = [
 ]
 
 const RESIDENT_DISCOVER_CONSUMERS_WINDOW = 24 * 60 * 60 // seconds
-const MAX_RESIDENT_DISCOVER_CONSUMERS_BY_WINDOW = 3
+const MAX_RESIDENT_DISCOVER_CONSUMERS_BY_WINDOW = 10
 
 module.exports = {
     PAYMENT_CATEGORIES_META,
